app/requirements.txt
celery==5.6.0
celery-types==0.23.0
fastapi==0.125.0
python-dotenv==1.2.1
redis==7.1.0
uvicorn==0.38.0
pydantic-settings==2.12.0
flower==2.0.1
